endo-fs: review pass D — extract wrap-backend modules + test cleanup
Addresses items 10 and 11 of designs/endo-fs-seam-review-followups.md.

Item 11 (module extraction): wrap-backend.js shrinks from 1306 to 872
lines by extracting the orthogonal building blocks into shared/:

- shared/qid.js — synthQid(path, kind) — pure FNV-1a hash for the
  inode-like pathId on backings without real inode identity.
- shared/stat-table.js — makeStatTable(lockKeyOf) → { statTable,
  touch, statOf, readStatNow } — vat-local mtime/atime/ctime/btime
  tracking with backend.getStat preference baked into readStatNow.
- shared/path-tables.js — cleanupPathTables / transplantPathTables
  — sweep/move Map<path-key, ...> tables on remove/rename. Used by
  wrap-backend for its stat / xattr / event-subscriber tables.
- shared/xattrs-exo.js — makeXattrsExo({ xattrTable, fireLocal,
  lockKeyOf, path }) — vat-local user.* xattr sidecar exo.
- shared/cursor-exo.js — makeCursorExo({ backend, dirPath }) —
  paged/streaming/drain Cursor over backend.list.
- shared/watcher-exo.js — makeNodeWatcherExo({ hasWatch,
  backendWatch, localSubs, subKey }) — merges backend.watch events
  with wrap-backend-local events; closes pump-cleanly on backend
  iterator error.

wrap-backend.js now binds the per-Filesystem closure values (backend,
caps, fireLocal, lockKeyOf, etc.) into thin local factories
(xattrsExoFor / cursorExoFor / watcherExoFor) that delegate to the
shared modules. Every call site replaced.

Item 10 (test cleanup):
- test/wrap-backend-fixes.test.js: @ts-check (was @ts-nocheck).
  Combined with the now-tighter interface guards, type drift in the
  new regression tests gets caught at lint time.
- "watcher unblocks pending consumers" test: drop the Promise.race
  with a manual setTimeout and rely on ava's t.timeout(2000) instead.
  Assert both that the first next() returns done AND that a second
  next() also returns done — the regression we're pinning is "the
  pump hangs forever," and asserting on stream-done semantics is
  more precise than racing a timer.

226 endo-fs tests + 20 9p-server tests pass; lint 0 errors; prettier
clean.
